This is an initial commit of DMSCMD, which is CMS command editing. See the HELP file for details.
It contains full source for DMSCMD, including updates, as well as the HISTORY EXEC that lets you
view and modify a full-screen list of saved commands.

File HISTORY MAILABLE is a self-extracting MAILABLE file: upload it as text, rename it to EXEC,
and run it--it will tell you what's in it and how to extract the files. It just duplicates the
rest of the files as a single blob, but also includes a DMSCMD MODULE so you don't have to
build it.

The other file that is not part of DMSCMD itself is HISTORY FILES. This is just a list of the
files in the MAILABLE, used with to create the MAILABLE via:
        MAILABLE HISTORY FILES fm HISTORY MAILABLE fm (FILELIST